Issue #841 move tab switcher to its own file and clean up
